+++ el-proyecte-grande-backend/el-proyecte-grande-backend.csproj
@@ -32,7 +32,6 @@
 
   <ItemGroup>
     <Folder Include="Models\Dtos\" />
-    <Folder Include="Repositories\" />
   </ItemGroup>
